requirements.txt
langchain == 0.1.19
langchain-community == 0.0.38
langchain-core == 0.1.52
langchain-experimental == 0.0.57
langchain-openai == 0.1.6
faiss-cpu == 1.8.0
pydantic == 1.10.9
pandas == 2.2.2
slackclient == 2.9.4
emoji == 2.11.1